Test Sub
Modeled a test submarine, loaded into Unity. It has a floater component
on it, but because of it's size the performance is pretty terrible with
floating, need to fix. Nothing else inside Unity has been implemented,
no colliders for walking in it or shaders or lighting or anything like
that.