A totally different approach to volume calculations.
now with things that track position in the water. This is a much simpler
(probably) approach, the only downside being that is not as fast a
direct calculation ie it takes several frames to get depth readings, so
fast moving ships may be artifically bouyant when sinking and heavy when
surfacing. Maybe that will work in my favor to balance out ridiculously
fast ships? add a curve for sink/rise speeds?