Floating Mostly Works
Floating with the new method now works for small objects, large objets
have an issue in which the tracker is falling at approximately the same
or slower speeds  as the sinking ship, and so the volume does not get
updated nearly fast enough. This leads to them sinking rapidly,
compounding the issue.