Yet Another Total Rewrite of the Floatation System
Becuase that's what I need. This one uses two hybrid tracker/rendered
things that both determine top and bottom height and display two
seperate parts of water so that they don't get in the boat. The
bottomWater script produces some weird ripplings on the bottom of the
water for unknown reasons, but players shouldn't see the bottom of the
ocean probably so it should be fine.